backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LlamaIndex and Hugging Face
llama-index==0.10.68
llama-index-llms-huggingface==0.3.5
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0

# Supabase
supabase==2.3.4

# Pinecone
pinecone-client==3.0.2

# Embeddings
sentence-transformers==2.3.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Additional utilities
python-dotenv==1.0.0
python-multipart==0.0.6
